+++ DigitalMenuApi/Controllers/BoxesController.cs
@@ -33,7 +33,7 @@ namespace DigitalMenuApi.Controllers
             var response = new PagingResponseDto<BoxReadDto> { Result = boxes, Count = dto.Count };
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit });
                 }
