+++ DigitalMenuApi/Controllers/StoreController.cs
@@ -48,7 +48,7 @@ namespace DigitalMenuApi.Controllers
             var response = new PagingResponseDto<StoreReadDto> { Result = store, Count = dto.Count };
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit, searchValue });
                 }
@@ -88,7 +88,7 @@ namespace DigitalMenuApi.Controllers
             var response = new PagingResponseDto<ProductReadDto> { Result = product, Count = dto.Count };
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit, searchValue });
                 }
@@ -114,7 +114,7 @@ namespace DigitalMenuApi.Controllers
 
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit, searchValue });
                 }
@@ -142,7 +142,7 @@ namespace DigitalMenuApi.Controllers
 
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit, tag, searchValue });
                 }
@@ -165,7 +165,7 @@ namespace DigitalMenuApi.Controllers
 
             if (limit > 0)
             {
-                if (dto.Count / limit > page)
+                if ((double)dto.Count / limit > page)
                 {
                     response.NextPage = Url.Link(null, new { page = page + 1, limit });
                 }
